feat and cleaning
added the args parser and implemented it to turn headless on and off easier. Should make making a gui easier as well.
fixed spelling. added a basic way of error handling if it doesnt grab the screenshot of a comment. still needs to be added on to so it doesnt just exit the program and can skip the comments when rendering the video.